requirements.txt
django>=4.0.0
mysqlclient>=2.2.0
pandas>=2.0.0
opencv-python
tensorflow
keras
seaborn
scikit-learn
plotly
scipy
streamlit
altair
protobuf
sklearn-evaluation
notebook
ipywidgets
matplotlib
mlxtend
numpy
